Updated pitBigTableBidsAndAsks

+++ src/microservices/warehouse/app/orm/models/pits/pit_bigtable_bids_and_asks.py
@@ -1,6 +1,6 @@
 from sqlalchemy import Column, Text, Float, text
 from clickhouse_sqlalchemy import engines
-from clickhouse_sqlalchemy.types.common import DateTime, UUID, Nullable
+from clickhouse_sqlalchemy.types.common import DateTime, UUID, Nullable, Int128, Int64, Float64
 from sqlalchemy.ext.declarative import declared_attr
 from fastapi_utils.camelcase import camel2snake
 
@@ -25,22 +25,22 @@ class pitBigTableBidsAndAsks(Base):
     h_network_name = Column(Text)
 
     pit_symbol = Column(Nullable(Text))
-    pit_reserve0 = Column(Nullable(Float))
-    pit_reserve1 = Column(Nullable(Float))
+    pit_reserve0 = Column(Nullable(Int128))
+    pit_reserve1 = Column(Nullable(Int128))
     pit_decimals0 = Column(Nullable(Float))
     pit_decimals1 = Column(Nullable(Float))
     pit_fee = Column(Nullable(Float))
-    pit_sqrt_p = Column(Nullable(Float))
-    pit_liquidity = Column(Nullable(Float))
+    pit_sqrt_p = Column(Nullable(Int128))
+    pit_liquidity = Column(Nullable(Int128))
     pit_price = Column(Nullable(Float))
     pit_sender = Column(Nullable(Text))
     pit_recipient = Column(Nullable(Text))
-    pit_amount0 = Column(Nullable(Float))
-    pit_amount1 = Column(Nullable(Float))
-    pit_gas_used = Column(Nullable(Float))
-    pit_effective_gas_price = Column(Nullable(Float))
+    pit_amount0 = Column(Nullable(Int128))
+    pit_amount1 = Column(Nullable(Int128))
+    pit_gas_used = Column(Nullable(Int128))
+    pit_effective_gas_price = Column(Nullable(Float64))
     pit_gas_symbol = Column(Nullable(Text))
-    pit_index_position_in_the_block = Column(Nullable(Float))
+    pit_index_position_in_the_block = Column(Nullable(Int64))
     pit_tx_hash = Column(Nullable(Text))
 
     pit_ts = Column(DateTime)
